DjangoTightrope/tightropeAPI/requirements.txt
asgiref==3.5.0
Django==4.0.1
django-auto-logout==0.5.0
django-cors-headers==3.11.0
djangorestframework==3.13.1
pytz==2021.3
sqlparse==0.4.2
